+++ packages/ragbits-document-search/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- BREAKING CHANGE: Renamed HttpSource to WebSource (#420)
+- BREAKING CHANGE: Renamed HttpSource to WebSource and changed property names (#420)
 - Better error distinction for WebSource (#421)
 
 ## 0.10.0 (2025-03-17)