+++ packages/ragbits-document-search/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - BREAKING CHANGE: Renamed HttpSource to WebSource (#420)
 - Better error distinction for WebSource (#421)
-
+- correct the typo in the class name ObjectConstructionConfig
+- 
 ## 0.10.0 (2025-03-17)
 
 ### Changed